+++ src/com/yetanalytics/dave/ui/app/workbook/data.cljs
@@ -279,6 +279,13 @@
                               :type :com.yetanalytics.dave.workbook.data/file
                               :uri "data/dave/ds.json"
                               :built-in? true}]}
+                 {:label "Kokea Concepts Test Dataset"
+                  :img-src "img/folder.png"
+                  :dispatch [::change workbook-id
+                             {:title "kokea test dataset"
+                              :type :com.yetanalytics.dave.workbook.data/file
+                              :uri "data/kokea/kokea.json"
+                              :built-in? true}]}
                  {:label "LRS Data"
                   :img-src "img/db.png"
                   :dispatch